+++ internal/connector/sink.go
@@ -34,4 +34,5 @@ type Sink interface {
 
 	InsertFileMetric(model.FileMetric) error
 	InsertHarnessArtifact(model.HarnessArtifact) error
+	InsertFileComplexityHistory(model.FileComplexityHistory) error
 }